requirements.txt
#Requirements.txt for dependencies
#  pip install -r requirements.txt

Adafruit-Blinka==3.3.4                    #
adafruit-circuitpython-busdevice==4.1.1   #
adafruit-circuitpython-framebuf==1.2.1    #
adafruit-circuitpython-ssd1306==2.7.0     #
Adafruit-GPIO==1.0.4                      #
Adafruit-PlatformDetect==1.3.8            #
Adafruit-PureIO==1.0.4                    #
Adafruit-SSD1306==1.6.2                   #
beautifulsoup4==4.7.1                     #
Flask==1.0.2                              #
folium==0.11.0                            #
gpiozero==1.5.1                           #
html5lib==1.0.1                           #
Jinja2==2.10                              #
logzero==1.5.0                            #
lxml==4.3.2                               #
numpy==1.16.2                             #
Pillow==5.4.1                             #
pysftp==0.2.9                             #
python-nmap==0.6.1                        #
requests==2.21.0                          #
rpi-ws281x==4.2.2                         #
RPi.GPIO==0.7.0                           #
RPLCD==1.2.2                              #
smbus2==0.3.0                             #
urllib3==1.24.1                           #
Werkzeug==0.14.1                          #
wget==3.2                                 #
